+++ src/kardy/widgets.nim
@@ -1,5 +1,4 @@
-from std/dom import value
-
+from std/dom import value,  OptionElement
 include pkg/karax/prelude
 
 import kardy/base
@@ -9,7 +8,7 @@ const selectTitleValue = "default"
 
 proc newCardSelect*(cards: seq[CardSetting]; title = "Select the Card"): VNode =
   result = buildHtml select:
-    option(selected = "", disabled = "", value = selectTitleValue): text title
+    option(selected = "", value = selectTitleValue): text title
     for card in cards:
       option(value = $card.id):
         text card.name
@@ -17,5 +16,5 @@ proc newCardSelect*(cards: seq[CardSetting]; title = "Select the Card"): VNode =
         result.value = ev.target.value
   result.value = selectTitleValue
 
-func unselect*(select: VNode) =
+proc unselect*(select: VNode) =
   select.value = selectTitleValue