+++ changelog.md
@@ -6,6 +6,7 @@
 - Fixed chromium support
 - Centered elements
 - Added action undo (discard and add probability)
+- Fixed select unselecting
 
 ## Version 0.2.0 (2023/10/01)